🔨 refactor(product): useProduct 커스텀 훅 분리
- 상품 관련 상태 및 로직을 useProduct 훅으로 모듈화
